Sprint 1 milestone 3

+++ E-Commerce/Controllers/ProductCategoriesController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using E_Commerce.Data;
 using E_Commerce.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace E_Commerce.Controllers
 {
@@ -19,6 +20,7 @@ namespace E_Commerce.Controllers
             _context = context;
         }
 
+        //[Authorize(Roles = "Administrator")]
         // GET: ProductCategories
         public async Task<IActionResult> Index()
         {
@@ -55,8 +57,12 @@ namespace E_Commerce.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> Create([Bind("Id,Name")] ProductCategory productCategory)
         {
+            if (!User.IsInRole("Administrator"))
+                return NotFound();
+
             if (ModelState.IsValid)
             {
                 _context.Add(productCategory);
@@ -67,8 +73,12 @@ namespace E_Commerce.Controllers
         }
 
         // GET: ProductCategories/Edit/5
+        [Authorize(Roles = "Editor")]
         public async Task<IActionResult> Edit(int? id)
         {
+            if (!User.IsInRole("Editor"))
+                return NotFound();
+
             if (id == null)
             {
                 return NotFound();
@@ -87,8 +97,13 @@ namespace E_Commerce.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Editor")]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] ProductCategory productCategory)
         {
+            if (!User.IsInRole("Editor"))
+                return NotFound();
+
+
             if (id != productCategory.Id)
             {
                 return NotFound();
@@ -118,8 +133,12 @@ namespace E_Commerce.Controllers
         }
 
         // GET: ProductCategories/Delete/5
+        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!User.IsInRole("Administrator"))
+                return NotFound();
+
             if (id == null)
             {
                 return NotFound();
@@ -138,8 +157,12 @@ namespace E_Commerce.Controllers
         // POST: ProductCategories/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!User.IsInRole("Administrator"))
+                return NotFound();
+
             var productCategory = await _context.ProductCategories.FindAsync(id);
             _context.ProductCategories.Remove(productCategory);
             await _context.SaveChangesAsync();
