+++ E-Commerce/Controllers/ProductsController.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using E_Commerce.Data;
 using E_Commerce.Models;
+using Microsoft.AspNetCore.Authorization;
 
 namespace E_Commerce.Controllers
 {
@@ -18,6 +19,8 @@ namespace E_Commerce.Controllers
         {
             _context = context;
         }
+
+        //[Authorize(Roles = "Administrator")]
 
         // GET: Products
         public async Task<IActionResult> Index()
@@ -46,8 +49,12 @@ namespace E_Commerce.Controllers
         }
 
         // GET: Products/Create
+        [Authorize(Roles = "Administrator")]
         public IActionResult Create()
         {
+            if (!User.IsInRole("Administrator"))
+                return NotFound();
+
             ViewData["ProductCategoryId"] = new SelectList(_context.ProductCategories, "Id", "Name");
             return View();
         }
@@ -57,8 +64,12 @@ namespace E_Commerce.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> Create([Bind("Id,Name,Price,InventoryAmount,Summary,Condition,ProductCategoryId")] Product product)
         {
+            if (!User.IsInRole("Administrator"))
+                return NotFound();
+
             if (ModelState.IsValid)
             {
                 _context.Add(product);
@@ -70,8 +81,14 @@ namespace E_Commerce.Controllers
         }
 
         // GET: Products/Edit/5
+        [Authorize(Roles = "Editor")]
         public async Task<IActionResult> Edit(int? id)
         {
+
+            if (!User.IsInRole("Editor"))
+                return NotFound();
+
+
             if (id == null)
             {
                 return NotFound();
@@ -91,8 +108,13 @@ namespace E_Commerce.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Editor")]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name,Price,InventoryAmount,Summary,Condition,ProductCategoryId")] Product product)
         {
+
+            if (!User.IsInRole("Editor"))
+                return NotFound();
+
             if (id != product.Id)
             {
                 return NotFound();
@@ -123,8 +145,12 @@ namespace E_Commerce.Controllers
         }
 
         // GET: Products/Delete/5
+        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!User.IsInRole("Administrator"))
+                return NotFound();
+
             if (id == null)
             {
                 return NotFound();
@@ -144,8 +170,12 @@ namespace E_Commerce.Controllers
         // POST: Products/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [Authorize(Roles = "Administrator")]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!User.IsInRole("Administrator"))
+                return NotFound();
+
             var product = await _context.Product.FindAsync(id);
             _context.Product.Remove(product);
             await _context.SaveChangesAsync();
